Changes to go along with bspace extension pr
