+++ netlogo-gui/resources/i18n/GUI_Strings_en.properties
@@ -671,6 +671,7 @@ tools.behaviorSpace.extension.noExperiment = No experiment exists with the name 
 tools.behaviorSpace.extension.guiExperiment = Experiment "{0}" is a GUI experiment, it cannot be edited.
 tools.behaviorSpace.extension.alreadyExists = An experiment already exists with the name "{0}".
 tools.behaviorSpace.extension.noReturn = No return value exists with the name "{0}".
+tools.behaviorSpace.extension.fileExists = File "{0}" already exists.
 
 tools.loggingMode.enterName = Enter your name:
 tools.loggingMode.fileDoesNotExist = The file {0} does not exist.
